leave_service/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
kafka-python
